// src/store/laboratories/laboratories.types.ts
export enum LABORATORY_ACTION_TYPES {
  GET_ALL_LABORATORY_START = 'laboratory/GET_ALL_LABORATORY_START',
  GET_ALL_LABORATORY_SUCCESS = 'laboratory/GET_ALL_LABORATORY_SUCCESS',
  GET_ALL_LABORATORY_FAILED = 'laboratory/GET_ALL_LABORATORY_FAILED',
  GET_LABORATORY_START = 'laboratory/GET_LABORATORY_START',
  GET_LABORATORY_SUCCESS = 'laboratory/GET_LABORATORY_SUCCESS',
  GET_LABORATORY_FAILED = 'laboratory/GET_LABORATORY_FAILED',
  GET_LIST_MEMBER_PENDING_START = 'laboratory/GET_LIST_MEMBER_PENDING_START',
  GET_LIST_MEMBER_PENDING_SUCCESS = 'laboratory/GET_LIST_MEMBER_PENDING_SUCCESS',
  GET_LIST_MEMBER_PENDING_FAILED = 'laboratory/GET_LIST_MEMBER_PENDING_SUCCESS',
  CREATE_LABORATORY_START = 'laboratory/CREATE_LABORATORY_START',
  CREATE_LABORATORY_SUCCESS = 'laboratory/CREATE_LABORATORY_SUCCESS',
  CREATE_LABORATORY_FAILED = 'laboratory/CREATE_LABORATORY_FAILED',
  REMOVE_LABORATORY_START = 'laboratory/REMOVE_LABORATORY_START',
  REMOVE_LABORATORY_SUCCESS = 'laboratory/REMOVE_LABORATORY_SUCCESS',
  REMOVE_LABORATORY_FAILED = 'laboratory/REMOVE_LABORATORY_FAILED',
};